:truck: Frontend module
