+++ modules/Backend/Actions/EnqueueStyleAction.php
@@ -19,7 +19,6 @@ class EnqueueStyleAction extends Action
     {
         $this->addAction(self::BACKEND_HEADER_ACTION, [$this, 'enqueueStylesHeader']);
         $this->addAction(self::BACKEND_FOOTER_ACTION, [$this, 'enqueueStylesFooter']);
-        $this->addAction(self::FRONTEND_HEADER_ACTION, [$this, 'addFrontendHeader']);
     }
 
     public function enqueueStylesHeader()
@@ -51,50 +50,6 @@ class EnqueueStyleAction extends Action
         foreach ($scripts as $script) {
             $href = e($script->get('src')) .'?v='. e($script->get('ver'));
             echo '<script src="'. $href .'"></script>';
-        }
-    }
-
-    public function addFrontendHeader()
-    {
-        $fbAppId = get_config('fb_app_id');
-        $googleAnalytics = get_config('google_analytics');
-        $scripts = HookAction::getEnqueueFrontendScripts();
-        $styles = HookAction::getEnqueueFrontendStyles();
-
-        echo e(
-            view(
-                'cms::items.frontend_header',
-                compact(
-                    'fbAppId',
-                    'googleAnalytics',
-                    'scripts',
-                    'styles'
-                )
-            )
-        );
-    }
-
-    public function addRecaptchaForm()
-    {
-        $this->addAction('auth_form', [$this, 'recaptchaRender']);
-    }
-
-    public function recaptchaRender()
-    {
-        $recaptcha = get_configs(
-            [
-                'google_recaptcha',
-                'google_recaptcha_key',
-            ]
-        );
-
-        if ($recaptcha['google_recaptcha'] == 1) {
-            echo view(
-                'cms::components.frontend.recaptcha',
-                compact(
-                    'recaptcha'
-                )
-            );
         }
     }
 }
