+++ modules/Backend/Helpers/helpers.php
@@ -3,7 +3,6 @@
 require __DIR__.'/html_dom.php';
 require __DIR__.'/data_helpers.php';
 require __DIR__.'/plugin.php';
-require __DIR__.'/theme.php';
 
 use Illuminate\Support\Arr;
 use Illuminate\Support\Carbon;
