+++ modules/Backend/Providers/BackendServiceProvider.php
@@ -5,29 +5,28 @@ namespace Juzaweb\Backend\Providers;
 use Illuminate\Foundation\AliasLoader;
 use Illuminate\Routing\Router;
 use Juzaweb\Backend\Actions\EnqueueStyleAction;
-use Juzaweb\Backend\Actions\FrontendAction;
 use Juzaweb\Backend\Actions\MenuAction;
 use Juzaweb\Backend\Actions\PermissionAction;
 use Juzaweb\Backend\Actions\SocialLoginAction;
 use Juzaweb\Backend\Actions\ThemeAction;
-use Juzaweb\Backend\Actions\ToolAction;
 use Juzaweb\Backend\Commands\EmailTemplateGenerateCommand;
 use Juzaweb\Backend\Commands\FindTransCommand;
 use Juzaweb\Backend\Commands\PermissionGenerateCommand;
 use Juzaweb\Backend\Commands\TransFromEnglish;
 use Juzaweb\Backend\Models\Comment;
+use Juzaweb\Backend\Models\Menu;
+use Juzaweb\Backend\Models\Post;
+use Juzaweb\Backend\Models\Taxonomy;
 use Juzaweb\Backend\Observers\CommentObserver;
+use Juzaweb\Backend\Observers\MenuObserver;
+use Juzaweb\Backend\Observers\PostObserver;
+use Juzaweb\Backend\Observers\TaxonomyObserver;
+use Juzaweb\CMS\Facades\ActionRegister;
 use Juzaweb\CMS\Http\Middleware\Admin;
 use Juzaweb\CMS\Support\Html\Field;
 use Juzaweb\CMS\Support\Macros\RouterMacros;
 use Juzaweb\CMS\Support\ServiceProvider;
-use Juzaweb\Backend\Models\Taxonomy;
-use Juzaweb\Backend\Observers\TaxonomyObserver;
-use Juzaweb\Backend\Models\Menu;
-use Juzaweb\Backend\Models\Post;
-use Juzaweb\Backend\Observers\MenuObserver;
-use Juzaweb\Backend\Observers\PostObserver;
-use Juzaweb\CMS\Facades\ActionRegister;
+use Juzaweb\Frontend\Actions\FrontendAction;
 
 class BackendServiceProvider extends ServiceProvider
 {
@@ -40,7 +39,7 @@ class BackendServiceProvider extends ServiceProvider
         Post::observe(PostObserver::class);
         Menu::observe(MenuObserver::class);
         Comment::observe(CommentObserver::class);
-    
+
         ActionRegister::register(
             [
                 MenuAction::class,
@@ -52,7 +51,7 @@ class BackendServiceProvider extends ServiceProvider
                 //ToolAction::class
             ]
         );
-    
+
         $this->commands(
             [
                 PermissionGenerateCommand::class,
@@ -89,14 +88,14 @@ class BackendServiceProvider extends ServiceProvider
     {
         $this->loadViewsFrom(__DIR__ . '/../resources/views', 'cms');
         $this->loadTranslationsFrom(__DIR__ . '/../resources/lang', 'cms');
-        
+
         $this->publishes(
             [
                 __DIR__ . '/../resources/views' => resource_path('views/vendor/juzaweb'),
             ],
             'cms_views'
         );
-        
+
         $this->publishes(
             [
                 __DIR__ . '/../resources/lang' => resource_path('lang/vendor/juzaweb'),
