+++ modules/CMS/Helpers/helpers.php
@@ -1,0 +1,5 @@
+<?php
+
+require_once __DIR__ . '/define.php';
+require_once __DIR__ . '/../../Backend/Helpers/helpers.php';
+require_once __DIR__ . '/../../Frontend/Helpers/helpers.php';
