+++ modules/Frontend/Helpers/helpers.php
@@ -1,0 +1,3 @@
+<?php
+
+require_once __DIR__ . '/theme.php';
