+++ modules/Frontend/Helpers/theme.php
