+++ modules/Frontend/Http/Controllers/PostController.php
@@ -8,16 +8,16 @@
  * @license    MIT
  */
 
-namespace Juzaweb\Backend\Http\Controllers\Frontend;
+namespace Juzaweb\Frontend\Http\Controllers;
 
-use Juzaweb\Backend\Events\PostViewed;
-use Juzaweb\Backend\Http\Resources\CommentResource;
-use Juzaweb\Backend\Models\Comment;
-use Juzaweb\CMS\Http\Controllers\FrontendController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Juzaweb\Backend\Events\PostViewed;
 use Juzaweb\Backend\Facades\HookAction;
+use Juzaweb\Backend\Http\Resources\CommentResource;
 use Juzaweb\Backend\Http\Resources\PostResource;
+use Juzaweb\Backend\Models\Comment;
+use Juzaweb\CMS\Http\Controllers\FrontendController;
 
 class PostController extends FrontendController
 {
