+++ modules/Frontend/routes/assets.php
